+++ src/main/java/de/amr/yt/pacman/ui/Sprites.java
@@ -56,8 +56,8 @@ public class Sprites {
 		return IT;
 	}
 
-	public static final int[] PACMAN_MOUTH_ANIMATION = { 1, 1, 0, 0, 1, 1, 2, 2 };
-	public static final int[] GHOST_ANIMATION = { 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1 };
+	public static final byte[] PACMAN_MOUTH_ANIMATION = { 1, 1, 0, 0, 1, 1, 2, 2 };
+	public static final byte[] GHOST_ANIMATION = { 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1 };
 
 	public BufferedImage mazeImage;
 	public BufferedImage sheetImage;
